requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
filetype==1.2.0
fonttools==4.54.1
fsspec==2024.10.0
idna==3.7
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
lapx==0.5.11
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.49
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.23
ultralytics-thop==2.0.9
urllib3==2.2.3
zipp==3.20.2
